✨ Remove unnecessary code from AddSingleStudent Widget
Removed unused imports and commented out code from the AddSingleStudent
widget in order to improve code readability and maintainability.
